% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon.shiny-package.R
\docType{package}
\name{loon.shiny-package}
\alias{loon.shiny}
\alias{loon.shiny-package}
\title{loon.shiny: Automatically Create a 'Shiny' App Based on Interactive 'Loon' Widgets}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Package 'shiny' provides interactive web applications in R. Package 'loon' is an interactive toolkit engaged in open-ended, creative and unscripted data exploration. The 'loon.shiny' package can take 'loon' widgets and display a selfsame 'shiny' app.
}
\author{
\strong{Maintainer}: Zehao Xu \email{z267xu@uwaterloo.ca}

Authors:
\itemize{
  \item R. Wayne Oldford \email{rwoldford@uwaterloo.ca}
}

}
\keyword{internal}
